pyproject.toml
[project]
name = "ai-llm-focus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "ollama>=0.5.1",
    "openai>=1.93.3",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "pip>=25.1.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
    "streamlit>=1.47.0",
    "torch>=2.7.1",
    "transformers>=4.53.2",
]
